+++ composer.json
@@ -16,54 +16,52 @@
     ],
     "require": {
         "php": ">=5.5.9",
-        "illuminate/auth": "5.1.*",
-        "illuminate/bus": "5.1.*",
-        "illuminate/broadcasting": "5.1.*",
-        "illuminate/config": "5.1.*",
-        "illuminate/container": "5.1.*",
-        "illuminate/contracts": "5.1.*",
-        "illuminate/cache": "5.1.*",
-        "illuminate/console": "5.1.*",
-        "illuminate/cookie": "5.1.*",
-        "illuminate/database": "5.1.*",
-        "illuminate/encryption": "5.1.*",
-        "illuminate/events": "5.1.*",
-        "illuminate/filesystem": "5.1.*",
-        "illuminate/hashing": "5.1.*",
-        "illuminate/http": "5.1.*",
-        "illuminate/pagination": "5.1.*",
-        "illuminate/queue": "5.1.*",
-        "illuminate/session": "5.1.*",
-        "illuminate/support": "5.1.*",
-        "illuminate/translation": "5.1.*",
-        "illuminate/validation": "5.1.*",
-        "illuminate/view": "5.1.*",
-        "monolog/monolog": "~1.0",
+        "illuminate/auth": "5.2.*",
+        "illuminate/broadcasting": "5.2.*",
+        "illuminate/bus": "5.2.*",
+        "illuminate/cache": "5.2.*",
+        "illuminate/config": "5.2.*",
+        "illuminate/container": "5.2.*",
+        "illuminate/contracts": "5.2.*",
+        "illuminate/database": "5.2.*",
+        "illuminate/encryption": "5.2.*",
+        "illuminate/events": "5.2.*",
+        "illuminate/filesystem": "5.2.*",
+        "illuminate/hashing": "5.2.*",
+        "illuminate/http": "5.2.*",
+        "illuminate/pagination": "5.2.*",
+        "illuminate/pipeline": "5.2.*",
+        "illuminate/queue": "5.2.*",
+        "illuminate/support": "5.2.*",
+        "illuminate/translation": "5.2.*",
+        "illuminate/validation": "5.2.*",
+        "monolog/monolog": "~1.11",
         "mtdowling/cron-expression": "~1.0",
-        "nikic/fast-route": "0.4.*",
+        "nikic/fast-route": "0.7.*",
         "paragonie/random_compat": "~1.1",
-        "symfony/dom-crawler": "2.7.*",
-        "symfony/http-kernel": "2.7.*",
-        "symfony/http-foundation": "2.7.*",
-        "symfony/var-dumper": "2.7.*"
+        "symfony/http-kernel": "2.8.*|3.0.*",
+        "symfony/http-foundation": "2.8.*|3.0.*"
     },
     "require-dev": {
         "phpunit/phpunit": "~4.0",
         "mockery/mockery": "~0.9"
     },
     "suggest": {
-        "vlucas/phpdotenv": "Required to use .env files (~1.0)."
+        "vlucas/phpdotenv": "Required to use .env files (~2.2)."
     },
     "autoload": {
         "psr-4": {
             "Laravel\\Lumen\\": "src/"
         },
-        "classmap": [
-            "src/Foundation"
-        ],
         "files": [
             "src/helpers.php"
         ]
     },
-    "minimum-stability": "dev"
+    "extra": {
+        "branch-alias": {
+            "dev-master": "5.2-dev"
+        }
+    },
+    "minimum-stability": "dev",
+    "prefer-stable": true
 }
