--- src/Console/Commands/stubs/DatabaseSeeder.stub
@@ -1,20 +1,0 @@
-<?php
-
-use Illuminate\Database\Seeder;
-use Illuminate\Database\Eloquent\Model;
-
-class DatabaseSeeder extends Seeder {
-
-    /**
-     * Run the database seeds.
-     *
-     * @return void
-     */
-    public function run()
-    {
-        Model::unguard();
-
-        // $this->call('UserTableSeeder');
-    }
-
-}
