--- src/Console/Commands/stubs/routes.stub
@@ -1,16 +1,0 @@
-<?php
-
-/*
-|--------------------------------------------------------------------------
-| Load The Cached Routes
-|--------------------------------------------------------------------------
-|
-| Here we will decode and unserialize the RouteCollection instance that
-| holds all of the route information for an application. This allows
-| us to instantaneously load the entire route map into the router.
-|
-*/
-
-app()->setRoutes(
-    unserialize(base64_decode('{{routes}}'))
-);
