--- src/Foundation/Testing/WithoutMiddleware.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace Illuminate\Foundation\Testing;
-
-use Exception;
-
-trait WithoutMiddleware
-{
-    /**
-     * @before
-     */
-    public function disableMiddlewareForAllTests()
-    {
-        if (method_exists($this, 'withoutMiddleware')) {
-            $this->withoutMiddleware();
-        } else {
-            throw new Exception('Unable to disable middleware. CrawlerTrait not used.');
-        }
-    }
-}
