+++ src/Testing/Concerns/MakesHttpRequests.php
@@ -1,0 +1,383 @@
+<?php
+
+namespace Laravel\Lumen\Testing\Concerns;
+
+use Illuminate\Support\Str;
+use Illuminate\Http\Request;
+
+trait MakesHttpRequests
+{
+    /**
+     * The last response returned by the application.
+     *
+     * @var \Illuminate\Http\Response
+     */
+    protected $response;
+
+    /**
+     * The current URL being viewed.
+     *
+     * @var string
+     */
+    protected $currentUri;
+
+    /**
+     * Visit the given URI with a JSON request.
+     *
+     * @param  string  $method
+     * @param  string  $uri
+     * @param  array  $data
+     * @param  array  $headers
+     * @return $this
+     */
+    public function json($method, $uri, array $data = [], array $headers = [])
+    {
+        $content = json_encode($data);
+
+        $headers = array_merge([
+            'CONTENT_LENGTH' => mb_strlen($content, '8bit'),
+            'CONTENT_TYPE' => 'application/json',
+            'Accept' => 'application/json',
+        ], $headers);
+
+        $this->call(
+            $method, $uri, [], [], [], $this->transformHeadersToServerVars($headers), $content
+        );
+
+        return $this;
+    }
+
+    /**
+     * Visit the given URI with a GET request.
+     *
+     * @param  string  $uri
+     * @param  array  $headers
+     * @return $this
+     */
+    public function get($uri, array $headers = [])
+    {
+        $server = $this->transformHeadersToServerVars($headers);
+
+        $this->call('GET', $uri, [], [], [], $server);
+
+        return $this;
+    }
+
+    /**
+     * Visit the given URI with a POST request.
+     *
+     * @param  string  $uri
+     * @param  array  $data
+     * @param  array  $headers
+     * @return $this
+     */
+    public function post($uri, array $data = [], array $headers = [])
+    {
+        $server = $this->transformHeadersToServerVars($headers);
+
+        $this->call('POST', $uri, $data, [], [], $server);
+
+        return $this;
+    }
+
+    /**
+     * Visit the given URI with a PUT request.
+     *
+     * @param  string  $uri
+     * @param  array  $data
+     * @param  array  $headers
+     * @return $this
+     */
+    public function put($uri, array $data = [], array $headers = [])
+    {
+        $server = $this->transformHeadersToServerVars($headers);
+
+        $this->call('PUT', $uri, $data, [], [], $server);
+
+        return $this;
+    }
+
+    /**
+     * Visit the given URI with a PATCH request.
+     *
+     * @param  string  $uri
+     * @param  array  $data
+     * @param  array  $headers
+     * @return $this
+     */
+    public function patch($uri, array $data = [], array $headers = [])
+    {
+        $server = $this->transformHeadersToServerVars($headers);
+
+        $this->call('PATCH', $uri, $data, [], [], $server);
+
+        return $this;
+    }
+
+    /**
+     * Visit the given URI with a DELETE request.
+     *
+     * @param  string  $uri
+     * @param  array  $data
+     * @param  array  $headers
+     * @return $this
+     */
+    public function delete($uri, array $data = [], array $headers = [])
+    {
+        $server = $this->transformHeadersToServerVars($headers);
+
+        $this->call('DELETE', $uri, $data, [], [], $server);
+
+        return $this;
+    }
+
+    /**
+     * Send the given request through the application.
+     *
+     * This method allows you to fully customize the entire Request object.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return $this
+     */
+    public function handle(Request $request)
+    {
+        $this->currentUri = $request->fullUrl();
+
+        $this->response = $this->app->prepareResponse($this->app->handle($request));
+
+        return $this;
+    }
+
+    /**
+     * Assert that the response contains JSON.
+     *
+     * @param  array|null  $data
+     * @return $this
+     */
+    protected function shouldReturnJson(array $data = null)
+    {
+        return $this->receiveJson($data);
+    }
+
+    /**
+     * Assert that the response contains JSON.
+     *
+     * @param  array|null  $data
+     * @return $this|null
+     */
+    protected function receiveJson($data = null)
+    {
+        return $this->seeJson($data);
+    }
+
+    /**
+     * Assert that the response contains an exact JSON array.
+     *
+     * @param  array  $data
+     * @return $this
+     */
+    public function seeJsonEquals(array $data)
+    {
+        $actual = json_encode(array_sort_recursive(
+            json_decode($this->response->getContent(), true)
+        ));
+
+        $this->assertEquals(json_encode(array_sort_recursive($data)), $actual);
+
+        return $this;
+    }
+
+    /**
+     * Assert that the response contains JSON.
+     *
+     * @param  array|null  $data
+     * @param  bool  $negate
+     * @return $this
+     */
+    public function seeJson(array $data = null, $negate = false)
+    {
+        if (is_null($data)) {
+            $this->assertJson(
+                $this->response->getContent(), "JSON was not returned from [{$this->currentUri}]."
+            );
+
+            return $this;
+        }
+
+        return $this->seeJsonContains($data, $negate);
+    }
+
+    /**
+     * Assert that the response doesn't contain JSON.
+     *
+     * @param  array|null  $data
+     * @return $this
+     */
+    public function dontSeeJson(array $data = null)
+    {
+        return $this->seeJson($data, true);
+    }
+
+    /**
+     * Assert that the JSON response has a given structure.
+     *
+     * @param  array|null  $structure
+     * @param  array|null  $responseData
+     * @return $this
+     */
+    public function seeJsonStructure(array $structure = null, $responseData = null)
+    {
+        if (is_null($structure)) {
+            return $this->seeJson();
+        }
+
+        if (! $responseData) {
+            $responseData = json_decode($this->response->getContent(), true);
+        }
+
+        foreach ($structure as $key => $value) {
+            if (is_array($value)) {
+                $this->assertArrayHasKey($key, $responseData);
+                $this->seeJsonStructure($structure[$key], $responseData[$key]);
+            } else {
+                $this->assertArrayHasKey($value, $responseData);
+            }
+        }
+    }
+
+    /**
+     * Assert that the response contains the given JSON.
+     *
+     * @param  array  $data
+     * @param  bool  $negate
+     * @return $this
+     */
+    protected function seeJsonContains(array $data, $negate = false)
+    {
+        $method = $negate ? 'assertFalse' : 'assertTrue';
+
+        $actual = json_decode($this->response->getContent(), true);
+
+        if (is_null($actual) || $actual === false) {
+            return $this->fail('Invalid JSON was returned from the route. Perhaps an exception was thrown?');
+        }
+
+        $actual = json_encode(array_sort_recursive(
+            (array) $actual
+        ));
+
+        foreach (array_sort_recursive($data) as $key => $value) {
+            $expected = $this->formatToExpectedJson($key, $value);
+
+            $this->{$method}(
+                Str::contains($actual, $expected),
+                ($negate ? 'Found unexpected' : 'Unable to find')." JSON fragment [{$expected}] within [{$actual}]."
+            );
+        }
+
+        return $this;
+    }
+
+    /**
+     * Format the given key and value into a JSON string for expectation checks.
+     *
+     * @param  string  $key
+     * @param  mixed  $value
+     * @return string
+     */
+    protected function formatToExpectedJson($key, $value)
+    {
+        $expected = json_encode([$key => $value]);
+
+        if (Str::startsWith($expected, '{')) {
+            $expected = substr($expected, 1);
+        }
+
+        if (Str::endsWith($expected, '}')) {
+            $expected = substr($expected, 0, -1);
+        }
+
+        return $expected;
+    }
+
+    /**
+     * Call the given URI and return the Response.
+     *
+     * @param  string  $method
+     * @param  string  $uri
+     * @param  array   $parameters
+     * @param  array   $cookies
+     * @param  array   $files
+     * @param  array   $server
+     * @param  string  $content
+     * @return \Illuminate\Http\Response
+     */
+    public function call($method, $uri, $parameters = [], $cookies = [], $files = [], $server = [], $content = null)
+    {
+        $this->currentUri = $this->prepareUrlForRequest($uri);
+
+        $request = Request::create(
+            $this->currentUri, $method, $parameters,
+            $cookies, $files, $server, $content
+        );
+
+        return $this->response = $this->app->prepareResponse(
+            $this->app->handle($request)
+        );
+    }
+
+    /**
+     * Turn the given URI into a fully qualified URL.
+     *
+     * @param  string  $uri
+     * @return string
+     */
+    protected function prepareUrlForRequest($uri)
+    {
+        if (Str::startsWith($uri, '/')) {
+            $uri = substr($uri, 1);
+        }
+
+        if (! Str::startsWith($uri, 'http')) {
+            $uri = $this->baseUrl.'/'.$uri;
+        }
+
+        return trim($uri, '/');
+    }
+
+    /**
+     * Transform headers array to array of $_SERVER vars with HTTP_* format.
+     *
+     * @param  array  $headers
+     * @return array
+     */
+    protected function transformHeadersToServerVars(array $headers)
+    {
+        $server = [];
+        $prefix = 'HTTP_';
+
+        foreach ($headers as $name => $value) {
+            $name = strtr(strtoupper($name), '-', '_');
+
+            if (! starts_with($name, $prefix) && $name != 'CONTENT_TYPE') {
+                $name = $prefix.$name;
+            }
+
+            $server[$name] = $value;
+        }
+
+        return $server;
+    }
+
+    /**
+     * Disable middleware for the test.
+     *
+     * @return $this
+     */
+    public function withoutMiddleware()
+    {
+        $this->app->instance('middleware.disable', true);
+
+        return $this;
+    }
+}
