+++ src/Testing/DatabaseTransactions.php
@@ -1,11 +1,13 @@
 <?php
 
-namespace Illuminate\Foundation\Testing;
+namespace Laravel\Lumen\Testing;
 
 trait DatabaseTransactions
 {
     /**
-     * @before
+     * Begin a database transaction.
+     *
+     * @return void
      */
     public function beginDatabaseTransaction()
     {
