+++ src/Testing/TestCase.php
@@ -4,17 +4,19 @@ namespace Laravel\Lumen\Testing;
 
 use Mockery;
 use PHPUnit_Framework_TestCase;
+use Illuminate\Support\Facades\Facade;
+use Illuminate\Contracts\Auth\Authenticatable;
 
 abstract class TestCase extends PHPUnit_Framework_TestCase
 {
-    use ApplicationTrait, AssertionsTrait, CrawlerTrait;
+    use Concerns\MakesHttpRequests;
 
     /**
-     * The callbacks that should be run before the application is destroyed.
+     * The application instance.
      *
-     * @var array
+     * @var \Laravel\Lumen\Application
      */
-    protected $beforeApplicationDestroyedCallbacks = [];
+    protected $app;
 
     /**
      * The base URL to use while testing the application.
@@ -22,6 +24,13 @@ abstract class TestCase extends PHPUnit_Framework_TestCase
      * @var string
      */
     protected $baseUrl = 'http://localhost';
+
+    /**
+     * The callbacks that should be run before the application is destroyed.
+     *
+     * @var array
+     */
+    protected $beforeApplicationDestroyedCallbacks = [];
 
     /**
      * Creates the application.
@@ -33,6 +42,20 @@ abstract class TestCase extends PHPUnit_Framework_TestCase
     abstract public function createApplication();
 
     /**
+     * Refresh the application instance.
+     *
+     * @return void
+     */
+    protected function refreshApplication()
+    {
+        putenv('APP_ENV=testing');
+
+        $this->app = $this->createApplication();
+
+        Facade::clearResolvedInstances();
+    }
+
+    /**
      * Setup the test environment.
      *
      * @return void
@@ -41,6 +64,34 @@ abstract class TestCase extends PHPUnit_Framework_TestCase
     {
         if (! $this->app) {
             $this->refreshApplication();
+        }
+
+        $this->setUpTraits();
+    }
+
+    /**
+     * Boot the testing helper traits.
+     *
+     * @return void
+     */
+    protected function setUpTraits()
+    {
+        $uses = array_flip(class_uses_recursive(get_class($this)));
+
+        if (isset($uses[DatabaseTransactions::class])) {
+            $this->beginDatabaseTransaction();
+        }
+
+        if (isset($uses[DatabaseMigrations::class])) {
+            $this->runDatabaseMigrations();
+        }
+
+        if (isset($uses[WithoutMiddleware::class])) {
+            $this->disableMiddlewareForAllTests();
+        }
+
+        if (isset($uses[WithoutEvents::class])) {
+            $this->disableEventsForAllTests();
         }
     }
 
@@ -63,6 +114,171 @@ abstract class TestCase extends PHPUnit_Framework_TestCase
             $this->app->flush();
             $this->app = null;
         }
+    }
+
+    /**
+     * Assert that a given where condition exists in the database.
+     *
+     * @param  string  $table
+     * @param  array  $data
+     * @return $this
+     */
+    protected function seeInDatabase($table, array $data)
+    {
+        $count = $this->app->make('db')->table($table)->where($data)->count();
+
+        $this->assertGreaterThan(0, $count, sprintf(
+            'Unable to find row in database table [%s] that matched attributes [%s].', $table, json_encode($data)
+        ));
+
+        return $this;
+    }
+
+    /**
+     * Assert that a given where condition does not exist in the database.
+     *
+     * @param  string  $table
+     * @param  array  $data
+     * @return $this
+     */
+    protected function missingFromDatabase($table, array $data)
+    {
+        return $this->notSeeInDatabase($table, $data);
+    }
+
+    /**
+     * Assert that a given where condition does not exist in the database.
+     *
+     * @param  string  $table
+     * @param  array  $data
+     * @return $this
+     */
+    protected function notSeeInDatabase($table, array $data)
+    {
+        $count = $this->app->make('db')->table($table)->where($data)->count();
+
+        $this->assertEquals(0, $count, sprintf(
+            'Found unexpected records in database table [%s] that matched attributes [%s].', $table, json_encode($data)
+        ));
+
+        return $this;
+    }
+
+    /**
+     * Specify a list of events that should be fired for the given operation.
+     *
+     * These events will be mocked, so that handlers will not actually be executed.
+     *
+     * @param  array|string  $events
+     * @return $this
+     */
+    public function expectsEvents($events)
+    {
+        $events = is_array($events) ? $events : func_get_args();
+
+        $mock = Mockery::spy('Illuminate\Contracts\Events\Dispatcher');
+
+        $mock->shouldReceive('fire')->andReturnUsing(function ($called) use (&$events) {
+            foreach ($events as $key => $event) {
+                if ((is_string($called) && $called === $event) ||
+                    (is_string($called) && is_subclass_of($called, $event)) ||
+                    (is_object($called) && $called instanceof $event)) {
+                    unset($events[$key]);
+                }
+            }
+        });
+
+        $this->beforeApplicationDestroyed(function () use (&$events) {
+            if ($events) {
+                throw new Exception(
+                    'The following events were not fired: ['.implode(', ', $events).']'
+                );
+            }
+        });
+
+        $this->app->instance('events', $mock);
+
+        return $this;
+    }
+
+    /**
+     * Mock the event dispatcher so all events are silenced.
+     *
+     * @return $this
+     */
+    protected function withoutEvents()
+    {
+        $mock = Mockery::mock('Illuminate\Contracts\Events\Dispatcher');
+
+        $mock->shouldReceive('fire');
+
+        $this->app->instance('events', $mock);
+
+        return $this;
+    }
+
+    /**
+     * Specify a list of jobs that should be dispatched for the given operation.
+     *
+     * These jobs will be mocked, so that handlers will not actually be executed.
+     *
+     * @param  array|string  $jobs
+     * @return $this
+     */
+    protected function expectsJobs($jobs)
+    {
+        $jobs = is_array($jobs) ? $jobs : func_get_args();
+
+        $mock = Mockery::mock('Illuminate\Bus\Dispatcher[dispatch]', [$this->app]);
+
+        foreach ($jobs as $job) {
+            $mock->shouldReceive('dispatch')->atLeast()->once()
+                ->with(Mockery::type($job));
+        }
+
+        $this->app->instance(
+            'Illuminate\Contracts\Bus\Dispatcher', $mock
+        );
+
+        return $this;
+    }
+
+    /**
+     * Set the currently logged in user for the application.
+     *
+     * @param  \Illuminate\Contracts\Auth\Authenticatable  $user
+     * @param  string|null  $driver
+     * @return $this
+     */
+    public function actingAs(Authenticatable $user, $driver = null)
+    {
+        $this->be($user, $driver);
+
+        return $this;
+    }
+
+    /**
+     * Set the currently logged in user for the application.
+     *
+     * @param  \Illuminate\Contracts\Auth\Authenticatable  $user
+     * @param  string|null  $driver
+     * @return void
+     */
+    public function be(Authenticatable $user, $driver = null)
+    {
+        $this->app['auth']->guard($driver)->setUser($user);
+    }
+
+    /**
+     * Call artisan command and return code.
+     *
+     * @param string  $command
+     * @param array   $parameters
+     * @return int
+     */
+    public function artisan($command, $parameters = [])
+    {
+        return $this->code = $this->app['Illuminate\Contracts\Console\Kernel']->call($command, $parameters);
     }
 
     /**
