+++ tests/FullApplicationTest.php
@@ -4,7 +4,7 @@ use Mockery as m;
 use Illuminate\Http\Request;
 use Laravel\Lumen\Application;
 
-class ExampleTest extends PHPUnit_Framework_TestCase
+class FullApplicationTest extends PHPUnit_Framework_TestCase
 {
     public function tearDown()
     {
@@ -77,40 +77,17 @@ class ExampleTest extends PHPUnit_Framework_TestCase
         $this->assertEquals('12', $response->getContent());
     }
 
-    public function testRequestToControllerWithParameters()
-    {
-        $app = new Application;
-
-        $app->get('/foo/{bar}', 'LumenTestController@actionWithParameter');
-
-        $response = $app->handle(Request::create('/foo/1', 'GET'));
-
-        $this->assertEquals(200, $response->getStatusCode());
-        $this->assertEquals('1', $response->getContent());
-    }
-
     public function testCallbackRouteWithDefaultParameter()
     {
         $app = new Application;
         $app->get('/foo-bar/{baz}', function ($baz = 'default-value') {
-            return response($baz);
+          return response($baz);
         });
 
         $response = $app->handle(Request::create('/foo-bar/something', 'GET'));
 
         $this->assertEquals(200, $response->getStatusCode());
         $this->assertEquals('something', $response->getContent());
-    }
-
-    public function testControllerRouteWithDefaultParameter()
-    {
-        $app = new Application;
-        $app->get('/foo-bar/{baz}', 'LumenTestController@actionWithDefaultParameter');
-
-        $response = $app->handle(Request::create('/foo-bar/something2', 'GET'));
-
-        $this->assertEquals(200, $response->getStatusCode());
-        $this->assertEquals('something2', $response->getContent());
     }
 
     public function testGlobalMiddleware()
@@ -133,7 +110,7 @@ class ExampleTest extends PHPUnit_Framework_TestCase
     {
         $app = new Application;
 
-        $app->routeMiddleware(['foo' => 'LumenTestMiddleware', 'passing' => 'LumenTestPlainMiddleware']);
+        $app->routeMiddleware(['foo' => 'LumenTestMiddleware']);
 
         $app->get('/', function () {
             return response('Hello World');
@@ -143,27 +120,11 @@ class ExampleTest extends PHPUnit_Framework_TestCase
             return response('Hello World');
         }]);
 
-        $app->get('/bar', ['middleware' => ['foo'], function () {
-            return response('Hello World');
-        }]);
-
-        $app->get('/fooBar', ['middleware' => 'passing|foo', function () {
-            return response('Hello World');
-        }]);
-
         $response = $app->handle(Request::create('/', 'GET'));
         $this->assertEquals(200, $response->getStatusCode());
         $this->assertEquals('Hello World', $response->getContent());
 
         $response = $app->handle(Request::create('/foo', 'GET'));
-        $this->assertEquals(200, $response->getStatusCode());
-        $this->assertEquals('Middleware', $response->getContent());
-
-        $response = $app->handle(Request::create('/bar', 'GET'));
-        $this->assertEquals(200, $response->getStatusCode());
-        $this->assertEquals('Middleware', $response->getContent());
-
-        $response = $app->handle(Request::create('/fooBar', 'GET'));
         $this->assertEquals(200, $response->getStatusCode());
         $this->assertEquals('Middleware', $response->getContent());
     }
@@ -188,9 +149,9 @@ class ExampleTest extends PHPUnit_Framework_TestCase
     {
         $app = new Application;
 
-        $app->routeMiddleware(['foo' => 'LumenTestParameterizedMiddleware', 'passing' => 'LumenTestPlainMiddleware']);
+        $app->routeMiddleware(['foo' => 'LumenTestParameterizedMiddleware']);
 
-        $app->get('/', ['middleware' => 'passing|foo:bar,boom', function () {
+        $app->get('/', ['middleware' => 'foo:bar,boom', function () {
             return response('Hello World');
         }]);
 
@@ -198,97 +159,6 @@ class ExampleTest extends PHPUnit_Framework_TestCase
 
         $this->assertEquals(200, $response->getStatusCode());
         $this->assertEquals('Middleware - bar - boom', $response->getContent());
-    }
-
-    public function testGroupRouteMiddleware()
-    {
-        $app = new Application;
-
-        $app->routeMiddleware(['foo' => 'LumenTestMiddleware', 'bar' => 'LumenTestMiddleware']);
-
-        $app->group(['middleware' => 'foo'], function ($app) {
-            $app->get('/', function () {
-                return 'Hello World';
-            });
-            $app->group([], function () {
-            });
-            $app->get('/fooBar', function () {
-                return 'Hello World';
-            });
-        });
-
-        $app->group(['middleware' => ['foo']], function ($app) {
-            $app->get('/fooFoo', function () {
-                return 'Hello World';
-            });
-        });
-
-        $app->group(['middleware' => 'bar|foo'], function ($app) {
-            $app->get('/fooFooBar', function () {
-                return 'Hello World';
-            });
-        });
-
-        $app->get('/foo', function () {
-            return 'Hello World';
-        });
-
-        $response = $app->handle(Request::create('/', 'GET'));
-        $this->assertEquals(200, $response->getStatusCode());
-        $this->assertEquals('Middleware', $response->getContent());
-
-        $response = $app->handle(Request::create('/foo', 'GET'));
-        $this->assertEquals(200, $response->getStatusCode());
-        $this->assertEquals('Hello World', $response->getContent());
-
-        $response = $app->handle(Request::create('/fooFoo', 'GET'));
-        $this->assertEquals(200, $response->getStatusCode());
-        $this->assertEquals('Middleware', $response->getContent());
-
-        $response = $app->handle(Request::create('/fooFooBar', 'GET'));
-        $this->assertEquals(200, $response->getStatusCode());
-        $this->assertEquals('Middleware', $response->getContent());
-
-        $response = $app->handle(Request::create('/fooBar', 'GET'));
-        $this->assertEquals(200, $response->getStatusCode());
-        $this->assertEquals('Middleware', $response->getContent());
-    }
-
-    public function testGroupRouteNestedMiddleware()
-    {
-        $app = new Application;
-
-        $app->routeMiddleware(['passing' => 'LumenTestPlainMiddleware', 'bar' => 'LumenTestMiddleware']);
-
-        $app->group(['middleware' => 'passing'], function ($app) {
-            $app->get('/foo', ['middleware' => 'bar', function () {
-                return 'Hello World';
-            }]);
-        });
-
-        $app->group(['middleware' => ['passing']], function ($app) {
-            $app->get('/bar', ['middleware' => ['bar'], function () {
-                return 'Hello World';
-            }]);
-        });
-
-        $app->group(['middleware' => ['passing']], function ($app) {
-            $app->get('/fooBar', ['middleware' => 'passing|bar', function () {
-                return 'Hello World';
-            }]);
-        });
-
-        $response = $app->handle(Request::create('/foo', 'GET'));
-        $this->assertEquals(200, $response->getStatusCode());
-        $this->assertEquals('Middleware', $response->getContent());
-
-        $response = $app->handle(Request::create('/bar', 'GET'));
-        $this->assertEquals(200, $response->getStatusCode());
-        $this->assertEquals('Middleware', $response->getContent());
-
-        $response = $app->handle(Request::create('/fooBar', 'GET'));
-        $this->assertEquals(200, $response->getStatusCode());
-        $this->assertEquals('Middleware', $response->getContent());
     }
 
     public function testWithMiddlewareDisabled()
@@ -306,37 +176,6 @@ class ExampleTest extends PHPUnit_Framework_TestCase
 
         $this->assertEquals(200, $response->getStatusCode());
         $this->assertEquals('Hello World', $response->getContent());
-    }
-
-    public function testGroupPrefixRoutes()
-    {
-        $app = new Application;
-
-        $app->group(['prefix' => 'user'], function ($app) {
-            $app->get('/', function () {
-                return response('User Index');
-            });
-
-            $app->get('profile', function () {
-                return response('User Profile');
-            });
-
-            $app->get('/show', function () {
-                return response('User Show');
-            });
-        });
-
-        $response = $app->handle(Request::create('/user', 'GET'));
-        $this->assertEquals(200, $response->getStatusCode());
-        $this->assertEquals('User Index', $response->getContent());
-
-        $response = $app->handle(Request::create('/user/profile', 'GET'));
-        $this->assertEquals(200, $response->getStatusCode());
-        $this->assertEquals('User Profile', $response->getContent());
-
-        $response = $app->handle(Request::create('/user/show', 'GET'));
-        $this->assertEquals(200, $response->getStatusCode());
-        $this->assertEquals('User Show', $response->getContent());
     }
 
     public function testNotFoundResponse()
@@ -367,41 +206,6 @@ class ExampleTest extends PHPUnit_Framework_TestCase
         $response = $app->handle(Request::create('/', 'GET'));
 
         $this->assertEquals(405, $response->getStatusCode());
-    }
-
-    public function testControllerResponse()
-    {
-        $app = new Application;
-
-        $app->get('/', 'LumenTestController@action');
-
-        $response = $app->handle(Request::create('/', 'GET'));
-
-        $this->assertEquals(200, $response->getStatusCode());
-        $this->assertEquals('LumenTestController', $response->getContent());
-    }
-
-    public function testNamespacedControllerResponse()
-    {
-        $app = new Application;
-
-        require_once __DIR__.'/fixtures/TestController.php';
-
-        $app->group(['namespace' => 'Lumen\Tests'], function ($app) {
-            $app->get('/', 'TestController@action');
-            $app->group([], function () {
-            });
-            $app->get('/foo', 'TestController@action');
-        });
-
-        $response = $app->handle(Request::create('/', 'GET'));
-
-        $this->assertEquals(200, $response->getStatusCode());
-        $this->assertEquals('Lumen\Tests\TestController', $response->getContent());
-
-        $response = $app->handle(Request::create('/foo', 'GET'));
-        $this->assertEquals(200, $response->getStatusCode());
-        $this->assertEquals('Lumen\Tests\TestController', $response->getContent());
     }
 
     public function testGeneratingUrls()
@@ -497,6 +301,45 @@ class ExampleTest extends PHPUnit_Framework_TestCase
         $this->assertEquals(true, $_SERVER['__middleware.response']);
     }
 
+    public function testBasicControllerDispatching()
+    {
+        $app = new Application;
+
+        $app->get('/show/{id}', 'LumenTestController@show');
+
+        $response = $app->handle(Request::create('/show/25', 'GET'));
+
+        $this->assertEquals(200, $response->getStatusCode());
+        $this->assertEquals('25', $response->getContent());
+    }
+
+    public function testBasicControllerDispatchingWithGroup()
+    {
+        $app = new Application;
+        $app->routeMiddleware(['test' => LumenTestMiddleware::class]);
+
+        $app->group(['middleware' => 'test'], function ($app) {
+            $app->get('/show/{id}', 'LumenTestController@show');
+        });
+
+        $response = $app->handle(Request::create('/show/25', 'GET'));
+
+        $this->assertEquals(200, $response->getStatusCode());
+        $this->assertEquals('Middleware', $response->getContent());
+    }
+
+    public function testBasicControllerDispatchingWithMiddlewareIntercept()
+    {
+        $app = new Application;
+        $app->routeMiddleware(['test' => LumenTestMiddleware::class]);
+        $app->get('/show/{id}', 'LumenTestControllerWithMiddleware@show');
+
+        $response = $app->handle(Request::create('/show/25', 'GET'));
+
+        $this->assertEquals(200, $response->getStatusCode());
+        $this->assertEquals('Middleware', $response->getContent());
+    }
+
     public function testEnvironmentDetection()
     {
         $app = new Application;
@@ -504,6 +347,19 @@ class ExampleTest extends PHPUnit_Framework_TestCase
         $this->assertEquals('production', $app->environment());
         $this->assertTrue($app->environment('production'));
         $this->assertTrue($app->environment(['production']));
+    }
+
+    public function testValidationHelpers()
+    {
+        $app = new Application;
+
+        $app->get('/', function (Illuminate\Http\Request $request) {
+            $this->validate($request, ['name' => 'required']);
+        });
+
+        $response = $app->handle(Request::create('/', 'GET'));
+
+        $this->assertEquals(422, $response->getStatusCode());
     }
 }
 
@@ -515,6 +371,32 @@ class LumenTestServiceProvider extends Illuminate\Support\ServiceProvider
 {
     public function register()
     {
+    }
+}
+
+class LumenTestController
+{
+    public function __construct(LumenTestService $service)
+    {
+        //
+    }
+
+    public function show($id)
+    {
+        return $id;
+    }
+}
+
+class LumenTestControllerWithMiddleware extends Laravel\Lumen\Routing\Controller
+{
+    public function __construct(LumenTestService $service)
+    {
+        $this->middleware('test');
+    }
+
+    public function show($id)
+    {
+        return $id;
     }
 }
 
@@ -542,30 +424,5 @@ class LumenTestParameterizedMiddleware
     public function handle($request, $next, $parameter1, $parameter2)
     {
         return response("Middleware - $parameter1 - $parameter2");
-    }
-}
-
-class LumenTestController
-{
-    public $service;
-
-    public function __construct(LumenTestService $service)
-    {
-        $this->service = $service;
-    }
-
-    public function action()
-    {
-        return response(__CLASS__);
-    }
-
-    public function actionWithParameter($baz)
-    {
-        return response($baz);
-    }
-
-    public function actionWithDefaultParameter($baz = 'default-value')
-    {
-        return response($baz);
     }
 }
