--- tests/fixtures/TestController.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace Lumen\Tests;
-
-class TestController
-{
-    public function action()
-    {
-        return __CLASS__;
-    }
-}
